+++ packages/core/src/modules/user/__test__/synced-user-service.e2e-spec.ts
@@ -7,15 +7,15 @@ import { describe, expect, it } from "vitest"
 import { mockDeep } from "vitest-mock-extended"
 import { mockAuth0UserResponse } from "../../../../mock"
 import { createServiceLayerForTesting } from "../../../../vitest-integration.setup"
+import { type Auth0Repository, Auth0RepositoryImpl } from "../../external/auth0-repository"
 import {
   type NotificationPermissionsRepository,
   NotificationPermissionsRepositoryImpl,
 } from "../notification-permissions-repository"
 import { type PrivacyPermissionsRepository, PrivacyPermissionsRepositoryImpl } from "../privacy-permissions-repository"
-import { type SyncedUserService, SyncedUserServiceImpl } from "../synced-user-service"
 import { type UserRepository, UserRepositoryImpl } from "../user-repository"
-import { type ReadOnlyUserService, UserServiceImpl } from "../user-service"
-import { type Auth0Repository, Auth0RepositoryImpl } from "../../external/auth0-repository"
+import { type UserService, UserServiceImpl } from "../user-service"
+import { Auth0UserSyncService, type UserSyncService } from "../user-sync-service"
 
 interface ServerLayerOptions {
   db: Kysely<Database>
@@ -30,13 +30,13 @@ const createServiceLayer = async ({ db, auth0MgmtClient }: ServerLayerOptions) =
   const notificationPermissionsRepository: NotificationPermissionsRepository =
     new NotificationPermissionsRepositoryImpl(db)
 
-  const userService: ReadOnlyUserService = new UserServiceImpl(
+  const userService: UserService = new UserServiceImpl(
     userRepository,
     privacyPermissionsRepository,
     notificationPermissionsRepository
   )
 
-  const syncedUserService: SyncedUserService = new SyncedUserServiceImpl(userService, auth0Repository)
+  const syncedUserService: UserSyncService = new Auth0UserSyncService(userService, auth0Repository)
 
   return {
     userService,
