+++ packages/core/src/modules/user/user-service.ts
@@ -11,8 +11,23 @@ import type { Cursor } from "../../utils/db-utils"
 import type { NotificationPermissionsRepository } from "./notification-permissions-repository"
 import type { PrivacyPermissionsRepository } from "./privacy-permissions-repository"
 import type { UserRepository } from "./user-repository"
+import type { UserSyncService } from "./user-sync-service"
 
-export interface ReadOnlyUserService {
+// Until we have gather this data from the user, this fake data is used as the initial data for new users
+const FAKE_USER_EXTRA_SIGNUP_DATA: Omit<UserWrite, "email" | "id" | "auth0Id"> = {
+  givenName: "firstName",
+  familyName: "lastName",
+  middleName: "middleName",
+  name: "firstName middleName lastName",
+  allergies: ["allergy1", "allergy2"],
+  picture: "https://example.com/image.jpg",
+  studyYear: -1,
+  lastSyncedAt: new Date(),
+  phone: "12345678",
+  gender: "male",
+}
+
+export interface UserService {
   getById(id: UserId): Promise<User | null>
   getAll(limit: number): Promise<User[]>
   getPrivacyPermissionsByUserId(id: string): Promise<PrivacyPermissions>
@@ -23,14 +38,34 @@ export interface ReadOnlyUserService {
   searchByFullName(searchQuery: string, take: number, cursor?: Cursor): Promise<User[]>
   create(data: UserWrite): Promise<User>
   getByAuth0Id(auth0Id: string): Promise<User | null>
+  update(user: User): Promise<User>
+  handlePopulateUserWithFakeData(auth0Id: string, email?: string | null): Promise<void>
 }
 
-export class UserServiceImpl implements ReadOnlyUserService {
+export class UserServiceImpl implements UserService {
   constructor(
     private readonly userRepository: UserRepository,
     private readonly privacyPermissionsRepository: PrivacyPermissionsRepository,
-    private readonly notificationPermissionsRepository: NotificationPermissionsRepository
+    private readonly notificationPermissionsRepository: NotificationPermissionsRepository,
+    private readonly userSyncService: UserSyncService
   ) {}
+
+  async handlePopulateUserWithFakeData(auth0Id: string, email?: string | null) {
+    if (!email) {
+      throw new Error("Did not get email in jwt")
+    }
+    const user = await this.create({
+      ...FAKE_USER_EXTRA_SIGNUP_DATA,
+      email: email,
+      auth0Id: auth0Id,
+    })
+
+    await this.userSyncService.update(user)
+  }
+
+  async update(user: User) {
+    return this.userSyncService.update(user)
+  }
 
   async getByAuth0Id(auth0Id: string) {
     return this.userRepository.getByAuth0Id(auth0Id)
